Testing arcanist/phabricator
Summary: as mentioned, testing arcanist/phabricator

Test Plan: check

Reviewers: cprodescu

Reviewed By: cprodescu

Differential Revision: http://phab.code4fun.de/D36

git-svn-id: https://svn.kwarc.info/repos/MMT@1943 09069c07-60fa-4895-8316-727242e252e4

+++ src/mmt-mizar/trunk/src/main/info/kwarc/mmt/mizar/test/MwsService.scala
@@ -107,7 +107,7 @@ class MwsService() extends QueryTransformer {
   def parseQuery(n : scala.xml.Node, aid : String, mmlversion : String) : scala.xml.Node = {
     TranslationController.query = true
     TranslationController.currentAid = aid
-    TranslationController.currentVersion = mmlversion.toInt
+    TranslationController.currentVersion = mmlversion
     
     var nr = n.child.length
     val qvars = n.child.slice(0, nr - 1)
